minor fix destination port

minor script change


improve sim


TCP - change rx commands numbers to be added to one another


minor


Revert "TCP - change rx commands numbers to be added to one another"

This reverts commit 3acb5a5bb38f01637a1de34d847b771dcc4c8864.

move previous fix to json


fix simolator


fix SFR profile


add profile


WIP


WIP


WIP


convert buffer size to 32bit


fix sim


rm scripts

+++ src/astf/json_reader.cpp
@@ -390,7 +390,7 @@ bool CJsonData::build_assoc_translation(uint8_t socket_id) {
 bool CJsonData::convert_bufs(uint8_t socket_id) {
     CMbufBuffer *tcp_buf;
     std::string json_buf;
-    uint16_t buf_len;
+    uint32_t buf_len;
 
     if (m_val["buf_list"].size() == 0)
         return false;
